+++ cost_center_read.go
@@ -165,9 +165,9 @@ type CostCenter struct {
 	TextFreeField3    string
 	TextFreeField4    string
 	TextFreeField5    string
-	NumberFreeField1  int
-	NumberFreeField2  int
-	NumberFreeField3  int
-	NumberFreeField4  int
-	NumberFreeField5  int
+	NumberFreeField1  float64
+	NumberFreeField2  float64
+	NumberFreeField3  float64
+	NumberFreeField4  float64
+	NumberFreeField5  float64
 }
